refactor: improve client codes using context managers
