+++ src/sotool/integrations/__init__.py
@@ -1,0 +1,10 @@
+from .excel import ExcelClient
+from .outook import OutlookClient
+from .sap_connector import SAPConnector
+
+
+__all__ = [
+    "ExcelClient",
+    "OutlookClient",
+    "SAPConnector",
+]
