+++ src/sotool/integrations/pdf_parser.py
@@ -1,4 +1,4 @@
-# pdf_processor.py
+# TODO - make it flexible to parse
 import pdfplumber
 import re
 from typing import List
