--- src/sotool/utils/__init__.py
@@ -1,15 +1,0 @@
-from .excel_client import ExcelClient
-from .outlook_client import OutlookClient
-from .sap_utils import SAPUtils
-from .excel_utils import apply_borders, format_number, get_df_from_excel
-
-
-__all__ = [
-    "format_number",
-    "OutlookClient",
-    "SAPUtils",
-    "ExcelClient",
-    "apply_borders",
-    "get_df_from_excel",
-    "format_number",
-]
